+++ app.js
@@ -3,13 +3,22 @@ const cp = require('child_process');
 const linter = require('./lib/linter');
 const upgrader = require('./lib/upgrader');
 const { stripIndent } = require('common-tags');
+const prompts = require('prompts');
 
 if (argv._[0] === 'reset') {
   cp.execSync('git reset --hard && git clean -df');
 } else if (argv._[0] === 'lint') {
   linter({ argv });
 } else if (argv._[0] === 'upgrade') {
-  upgrader({ argv });
+  confirm()
+    .then(response => {
+      if (response.proceed) {
+        upgrader({ argv });
+      }
+    })
+    .catch(err => {
+      console.error(err);
+    });
 } else if (argv._[0] === 'help' || argv.help) {
   console.log(stripIndent`
     Commands:
@@ -28,4 +37,16 @@ if (argv._[0] === 'reset') {
   console.log(stripIndent`
     Run \`apos-code-upgrader help\` or \`apos-code-upgrader --help\` for commands.
 `);
+}
+
+async function confirm() {
+  return prompts({
+    type: 'confirm',
+    name: 'proceed',
+    message: stripIndent`
+      Running the upgrade command will make MAJOR changes to your code. It will not (git) commit those changes, however.
+
+      Are you sure you want to continue?`,
+    initial: false
+  });
 }