+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Adds a prompt before the upgrade script runs to confirm the user wants to proceed.
+
 ## 1.0.0-alpha 2021-11-23
 
 Initial alpha release.